feat: implement multiplexer app for tx settlement chaining
Overcomes a framework’s limitation, which is: chaining a skill more than once.

The limitation appears because we need to perform a refactor for sending a safe tx for performing a request to the mech instead of using the mech client, because the latter requires the ethereum private key's path.

Therefore, will need to send a safe tx both for a) performing a request to the mech and b) placing a bet.