fix: remove abci connection override
Remove the abci connection's override from the service file because it was not allowing any other connection's override to work properly. The override is either way unnecessary. The only reason it was being overriden is that the service analyzer (`autonomy analyse service --public-id valory/trader`) otherwise warns:

```
Overrides with following component IDs are defined in the agent configuration but not in the service configuration
        - (connection, valory/abci:0.1.0)
```