refactor: set the bets on fetch error
That is to avoid an issue in which the bets have never been set, because of a fetch error, and then we transition to a new period. The bets are in the cross-period keys, which means that the framework will attempt to access them, but they will never have been defined.