fix: payout based on subgraph utilization
This fixes an issue introduced in #172.

When we get the payout from the events on-chain, `0` is returned for the positions which have already been redeemed. However, this is not the case when we retrieve the payout from the subgraph (`get_position_lifetime_value` calculates the lifetime value returned from the conditional tokens' subgraph). In the latter case, all the payout amounts are returned, regardless of whether we have redeemed or not.

Therefore, when we use the on-chain events, we need to sum the current payouts with the payouts so far. However, when we use the subgraph, we need to use only the currently calculated payouts.