fix: issue with cross-period keys and the benchmarking mode
Fixes the issue with cross-period persisted keys not being available after a reset when using the benchmarking mode.

This change should ideally be moved to the core to avoid hacking around here. Extending the `setup` method with this functionality can be the solution in https://github.com/valory-xyz/open-autonomy/issues/2131.