fix: do not blacklist if the tool selection has not been run
This fixes an edge case which is caused when the tx settlement round fails and no tool has been selected. This can happen in two cases: a) after the nevermined subscription round b) after the KPI has been met and the tool selection is skipped intentionally. If this happens, the `HandleFailedTxRound` is called, and then the `BlacklistingRound`. During the blacklisting, no tool has been selected for the current round, and therefore we get:

```
ValueError: 'mech_tool_idx' field is not set for this period [x] and no default value was provided.
```