feat: update the bets' logic
The logic was updated to allow for multiple bets on the same market. The bet status is not necessary anymore. It is replaced by the timestamp in which we processed a bet (either bet on it or blacklisted it) and the number of bets we have placed on it. From now on, we can check when we last placed a bet on a market to decide whether we can bet on it or not, and we can also check whether the number of bets has been increased to determine whether the action taken was to bet or to blacklist it.