fix: correctly determine whether a new epoch is reached
Previously, we were only aware of a new epoch if our own agent was the one to call the checkpoint. If any other agent had called it, then we would never be aware of this event.

Instead, we are now tracking changes to the ts checkpoint of the contract which only and always gets updated after a new epoch is triggered.