fix: take into account the irrelevant tools
Fixes an issue introduced in https://github.com/valory-xyz/trader/pull/343. The irrelevant tools would not be considered unless there was a replacement of the local with the global information.

After this commit, the irrelevant tools are always considered and only the accuracy is not adjusted in case the trader decides to keep the local over the global information.