feat: get the metadata based on the mech version
The legacy mechs can access the metadata using the agent registry contract. However, the new mech marketplace can get them by using the complementary service metadata contract. The former method requires the mech's id as an input, but the latter requires the service id as an input. The mech id can be obtained from the legacy mech contract, but the service id can be obtained from the new mech contract, the one for the mech marketplace.