fix: multi-bets issue
if a bet was in the `TO_PROCESS` queue and could not be selected because of the constraints (e.g., not in opening margin), then everything was blocked because the `FRESH` status would never be updated.

This is fixed in this commit by blacklisting the unprocessable bets of the `TO_PROCESS` queue.