fix: remove unnecessary RPC calls
Double-checking whether a market is finalized is unnecessary, plus a conflict has not been found yet in the logs.

Therefore, removing it is sensible, especially taking into consideration the amount in which it affects the number of RPC calls, therefore providers' subscription costs. For each trade a service has performed, a call will be made at each period. A period runs every 2-5 minutes, or even 1 minute if the KPI goals have been met. This can easily explode to thousands of extra requests per hour for the average service - especially in combination with https://github.com/valory-xyz/open-autonomy/issues/2368.